fix a spurious assertion
This checks that the types of a cloned repository is equal. Whereas, the
previous behavior of clone() and create() were "make a bare repository." It's
possible to go from working copy to bare in both Git and Mercurial, and this
causes issues in tests.

Making the flag assume wc=False ensures the working copy API will never be
active unless asked for explicitly.